[WIP] converted vectors.unlabelled_percentage to int

+++ thesisgenerator/utils/db.py
@@ -16,7 +16,7 @@ db.connect()
 class Vectors(pw.Model):
     algorithm = pw.CharField(null=False)  # how the *unigram* vectors were built, e.g. count_windows, Turian, word2vec
     dimensionality = pw.IntegerField(null=True)  # 0 to indicate no SVD was done, -1 for not applicable
-    unlabelled_percentage = pw.FloatField(default=100., null=True)  # how much of the unlabelled data was used
+    unlabelled_percentage = pw.IntegerField(default=100, null=True)  # how much of the unlabelled data was used
     unlabelled = pw.CharField(null=True)  # path to unlabelled corpus that data was used, if I did it
     path = pw.CharField(null=True)  # where on disk the vectors are stored
     composer = pw.CharField()  # what composer was used to build phrasal vectors (if any)
